+++ src/main/java/rbasamoyai/createbigcannons/crafting/casting/InvalidCastingError.java
@@ -5,24 +5,30 @@ import com.mojang.serialization.codecs.RecordCodecBuilder;
 
 import net.minecraft.Util;
 import net.minecraft.core.BlockPos;
+import net.minecraft.core.Holder;
 import net.minecraft.network.chat.Component;
 import net.minecraft.network.chat.MutableComponent;
-import net.minecraft.world.level.material.Fluid;
+import net.neoforged.neoforge.fluids.FluidStack;
 import rbasamoyai.createbigcannons.base.CBCRegistries;
-import rbasamoyai.createbigcannons.utils.CBCRegistryUtils;
+import rbasamoyai.createbigcannons.munitions.big_cannon.fluid_shell.EndFluidStack;
 
-public record InvalidCastingError(BlockPos pos, Fluid fluid, CannonCastShape shape) {
+public record InvalidCastingError(BlockPos pos, EndFluidStack fluidStack, CannonCastShape shape) {
 
     public static final MapCodec<InvalidCastingError> CODEC = RecordCodecBuilder.mapCodec(instance -> instance.group(
         BlockPos.CODEC.fieldOf("Position").forGetter(InvalidCastingError::pos),
-        CBCRegistryUtils.getFluidRegistry().byNameCodec().fieldOf("Fluid").forGetter(InvalidCastingError::fluid),
+        EndFluidStack.CODEC.fieldOf("Fluid").forGetter(InvalidCastingError::fluidStack),
         CBCRegistries.cannonCastShapes().byNameCodec().fieldOf("CastShape").forGetter(InvalidCastingError::shape)
     ).apply(instance, InvalidCastingError::new));
 
 	public MutableComponent getMessage() {
-		MutableComponent fluidText = Component.translatable(Util.makeDescriptionId("fluid", CBCRegistryUtils.getFluidLocation(this.fluid)));
+		MutableComponent fluidText = getFluidStackText(this.fluidStack);
 		MutableComponent shapeText = Component.translatable(Util.makeDescriptionId("cast_shape", CBCRegistries.cannonCastShapes().getKey(this.shape)));
 		return Component.translatable("exception.createbigcannons.casting", this.pos.getX(), this.pos.getY(), this.pos.getZ(), fluidText, shapeText);
 	}
+
+    public static MutableComponent getFluidStackText(EndFluidStack efstack) {
+        FluidStack stack = efstack.isEmpty() ? FluidStack.EMPTY : new FluidStack(Holder.direct(efstack.fluid()), efstack.amount(), efstack.components().asPatch());
+        return stack.getHoverName().copy();
+    }
 
 }