+++ CHANGELOG.md
@@ -48,6 +48,7 @@ Fixed:
 - Fixed honey bucket and chocolate buckets not showing up in creative tab
 - Fixed melting basin ponder category and pump orientation
 - Fixed sliding breech and quick-firing breech breechblock rendering
+- Fixed cannon cast error fluid name display
 
 ## [5.10.2] - 2026-01-17